Split Grid Layer block into 4 blocks based on template.
