+++ assets/css/amp-editor-story-blocks.css
@@ -53,7 +53,7 @@ div[data-type="amp/amp-story-page"] .editor-inner-blocks .editor-block-list__lay
 	background-color: initial;
 }
 
-.editor-block-list__layout div[data-type="amp/amp-story-grid-layer"] {
+.editor-block-list__layout div[data-amp-type="grid"] {
 	position: absolute;
 	height: 100%;
 	width: 100%;
@@ -137,28 +137,28 @@ div[data-amp-position="lower-third"] {
 	bottom: 0;
 }
 
-div[data-type="amp/amp-story-grid-layer"] .block-library-image__resizer {
+div[data-amp-type="grid"] .block-library-image__resizer {
 	max-width: 318px !important;
 	max-height: 533px !important;
 }
 
-.editor-block-list__layout div[data-type="amp/amp-story-grid-layer"].is-selected,
-.editor-block-list__layout div[data-type="amp/amp-story-grid-layer"].is-selected .editor-block-list__block {
+.editor-block-list__layout div[data-amp-type="grid"].is-selected,
+.editor-block-list__layout div[data-amp-type="grid"].is-selected .editor-block-list__block {
 	outline: 1px dotted #99c2e0;
 }
 
-.editor-block-list__layout div[data-type="amp/amp-story-grid-layer"]:not(.is-selected):not(.is-selected-parent) .editor-block-list__block {
+.editor-block-list__layout div[data-amp-type="grid"]:not(.is-selected):not(.is-selected-parent) .editor-block-list__block {
 	z-index: 0;
 }
 
-.editor-block-list__layout div[data-type="amp/amp-story-grid-layer"]:not(.is-selected):not(.is-selected-parent) .editor-block-list__block p[data-is-placeholder-visible="true"] ~ p {
+.editor-block-list__layout div[data-amp-type="grid"]:not(.is-selected):not(.is-selected-parent) .editor-block-list__block p[data-is-placeholder-visible="true"] ~ p {
 	display: none;
 }
 
-.editor-block-list__layout div[data-type="amp/amp-story-grid-layer"].is-selected ~ div[data-type="amp/amp-story-grid-layer"],
-.editor-block-list__layout div[data-type="amp/amp-story-grid-layer"].is-selected-parent ~ div[data-type="amp/amp-story-grid-layer"],
-.editor-block-list__layout div[data-type="amp/amp-story-grid-layer"].is-selected ~ div[data-type="amp/amp-story-cta-layer"],
-.editor-block-list__layout div[data-type="amp/amp-story-grid-layer"].is-selected-parent ~ div[data-type="amp/amp-story-cta-layer"]{
+.editor-block-list__layout div[data-amp-type="grid"].is-selected ~ div[data-amp-type="grid"],
+.editor-block-list__layout div[data-amp-type="grid"].is-selected-parent ~ div[data-amp-type="grid"],
+.editor-block-list__layout div[data-amp-type="grid"].is-selected ~ div[data-amp-type="grid"],
+.editor-block-list__layout div[data-amp-type="grid"].is-selected-parent ~ div[data-type="amp/amp-story-cta-layer"]{
 	display: none;
 }
