+++ blocks/amp-story/index.js
@@ -1,3 +1,5 @@
 import './amp-story-page';
 import './amp-story-cta-layer';
-import './amp-story-grid-layer';
+import './amp-story-grid-layer-vertical';
+import './amp-story-grid-layer-thirds';
+import './amp-story-grid-layer-fill';
